Bugfix for ResourceAllocation
